Make the Rubocop::Minitest::Test suite re-usable by gem consumers
