+++ changelog/new_make_test_class_reusable_by_consumers.md
@@ -1,0 +1,1 @@
+* [#278](https://github.com/rubocop/rubocop-minitest/pull/278): Allow `RuboCop::Minitest::Test` to be used by consumers of the gem to test custom cops. ([@dougedey][])
