+++ test/rubocop/cop/minitest/assert_equal_test.rb
@@ -2,7 +2,7 @@
 
 require_relative '../../../test_helper'
 
-class AssertEqualTest < Minitest::Test
+class AssertEqualTest < RuboCop::Minitest::Test
   def test_registers_offense_when_using_assert_equal_operator_with_string
     assert_offense(<<~RUBY)
       class FooTest < Minitest::Test
