+++ test/rubocop/cop/minitest/assert_in_delta_test.rb
@@ -2,7 +2,7 @@
 
 require_relative '../../../test_helper'
 
-class AssertInDeltaTest < Minitest::Test
+class AssertInDeltaTest < RuboCop::Minitest::Test
   def test_registers_offense_when_using_assert_equal_with_float_as_expected_value
     assert_offense(<<~RUBY)
       class FooTest < Minitest::Test
