+++ test/rubocop/cop/minitest/assert_includes_test.rb
@@ -2,7 +2,7 @@
 
 require_relative '../../../test_helper'
 
-class AssertIncludesTest < Minitest::Test
+class AssertIncludesTest < RuboCop::Minitest::Test
   def test_registers_offense_when_using_assert_with_include
     assert_offense(<<~RUBY)
       class FooTest < Minitest::Test
