+++ test/rubocop/cop/minitest/assert_operator_test.rb
@@ -2,7 +2,7 @@
 
 require_relative '../../../test_helper'
 
-class AssertOperatorTest < Minitest::Test
+class AssertOperatorTest < RuboCop::Minitest::Test
   def test_registers_offense_when_using_assert_with_operator_method
     assert_offense(<<~RUBY)
       class FooTest < Minitest::Test
