+++ test/rubocop/cop/minitest/assert_truthy_test.rb
@@ -2,7 +2,7 @@
 
 require_relative '../../../test_helper'
 
-class AssertTruthyTest < Minitest::Test
+class AssertTruthyTest < RuboCop::Minitest::Test
   def test_registers_offense_when_using_assert_equal_with_true
     assert_offense(<<~RUBY)
       class FooTest < Minitest::Test
