+++ test/rubocop/cop/minitest/assertion_in_lifecycle_hook_test.rb
@@ -2,7 +2,7 @@
 
 require_relative '../../../test_helper'
 
-class AssertionInLifecycleHookTest < Minitest::Test
+class AssertionInLifecycleHookTest < RuboCop::Minitest::Test
   def test_registers_offense_when_using_bad_method
     assert_offense(<<~RUBY)
       class FooTest < Minitest::Test
