+++ test/rubocop/cop/minitest/duplicate_test_run_test.rb
@@ -2,7 +2,7 @@
 
 require_relative '../../../test_helper'
 
-class DuplicateTestRunTest < Minitest::Test
+class DuplicateTestRunTest < RuboCop::Minitest::Test
   def test_registers_offense_when_parent_and_child_have_tests_methods
     assert_offense(<<~RUBY)
       class ParentTest < Minitest::Test
