+++ test/rubocop/cop/minitest/refute_false_test.rb
@@ -2,7 +2,7 @@
 
 require_relative '../../../test_helper'
 
-class RefuteFalseTest < Minitest::Test
+class RefuteFalseTest < RuboCop::Minitest::Test
   def test_registers_offense_when_using_assert_equal_with_false
     assert_offense(<<~RUBY)
       class FooTest < Minitest::Test
