+++ test/rubocop/cop/minitest/refute_nil_test.rb
@@ -2,7 +2,7 @@
 
 require_relative '../../../test_helper'
 
-class RefuteNilTest < Minitest::Test
+class RefuteNilTest < RuboCop::Minitest::Test
   def test_registers_offense_when_using_refute_equal_with_nil
     assert_offense(<<~RUBY)
       class FooTest < Minitest::Test
