+++ test/rubocop/cop/minitest/refute_respond_to_test.rb
@@ -2,7 +2,7 @@
 
 require_relative '../../../test_helper'
 
-class RefuteRespondToTest < Minitest::Test
+class RefuteRespondToTest < RuboCop::Minitest::Test
   def test_registers_offense_when_using_refute_calling_respond_to_method
     assert_offense(<<~RUBY)
       class FooTest < Minitest::Test
