+++ test/rubocop/cop/minitest/unspecified_exception_test.rb
@@ -2,7 +2,7 @@
 
 require_relative '../../../test_helper'
 
-class UnspecifiedExceptionTest < Minitest::Test
+class UnspecifiedExceptionTest < RuboCop::Minitest::Test
   def test_registers_offense_when_using_assert_raises_without_exception
     assert_offense(<<~RUBY)
       class FooTest < Minitest::Test
