+++ test/test_helper.rb
@@ -5,6 +5,8 @@ require_relative '../lib/rubocop-minitest'
 # Require supporting files exposed for testing.
 require_relative '../lib/rubocop/minitest/support'
 require 'minitest/proveit'
+require 'minitest/pride'
+require 'minitest/autorun'
 
 Minitest::Test.class_eval do
   prove_it!
